requirements.txt
streamlit
langchain
langchain-google-genai
langchain-openai
langgraph
deepagents
langchain-ollama
python-dotenv